requirements.txt
# J.A.R.V.I.S. Dependencies
# Core system requirements

# System monitoring and control
psutil>=6.0.0
pywin32>=306

# Voice interface
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# GUI and interface
PyQt6>=6.7.1
PyQt6-Qt6>=6.7.1
qt6-tools>=6.5.0.1.3

# Image processing and computer vision
opencv-python>=4.10.0.84
Pillow>=10.4.0
pytesseract>=0.3.13
mss>=9.0.1

# Network and web
requests>=2.32.3
websocket-client>=1.8.0
selenium>=4.23.0
paho-mqtt>=2.1.0

# Data processing and storage
numpy>=2.1.1
pandas>=2.2.2
sqlalchemy>=2.0.35
rarfile>=4.2
nltk>=3.9.1

# Encryption and security
cryptography>=43.0.1
pyopenssl>=24.2.1

# Advanced features
torch>=2.4.1
transformers>=4.44.2
sentence-transformers>=3.1.1
chromadb>=0.5.5
tf-keras>=2.17.0
GPUtil>=1.4.0
azure-cognitiveservices-speech>=1.40.0

# Development and testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
nltk>=3.9.1

# Utilities
colorama>=0.4.6
tqdm>=4.66.5
pyyaml>=6.0.2
speedtest-cli>=2.1.3
pyserial>=3.5

# Optional but recommended
screeninfo>=0.8.1
winshell>=0.6.0
pypiwin32>=223